requirements.txt
apache-airflow-providers-google
google-cloud-bigquery
google-cloud-storage
pandas
